refactor(remote): remove legacy apiproxy protocol
按 RPC_ADAPTATION_PLAN.md 第 7 步删除旧协议传输层：

- 删除 harnessClient（旧点号 endpoint、/api/events.mux|events.host、
  /api/respond 与 SSE 读循环）、harnessState（旧双流协调器）、
  harnessConnection（旧连接代次）、harnessProtocol（旧 RpcMethod 表）。
- 随模块删除其钉住测试 harnessClient.test.js / harnessConnection.test.js。
- 仅存的三类领域类型（HarnessHostDescription/HarnessGoalEditChanges/
  HarnessQueueAction）与 store 记录 envelope HarnessStreamEnvelope 迁入
  types.ts；hostState/dshRuntime/sessionStore/sessionCatalog 改从 types 导入。
- store 的 applyMuxEnvelope/applyHostEnvelope reducer 保留：不再有生产
  调用方，但仍是护栏测试钉住的 store mutation 入口。

§12 静态检查通过：src/ 与 webview/ 无 events.mux、events.host、
/api/respond、session.history、workspace.list 与旧点号 endpoint；/api/
字符串仅存于 remote carrier。npm run check、compile、npm test（50/50）全绿。
